backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "anyio>=4.9.0,<4.12",
    "fastapi>=0.128.4",
    "greenlet>=3.3.1",
    "jose>=1.0.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "qrcode>=8.2",
    "redis>=7.1.1",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
    "email-validator>=2.3.0",
    "bcrypt>=5.0.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
